db: allow passing multiple -r options
Use vector to define the list of arguments repositories to be opened
by the pkgdb_* function, to keep ABI and API compatibility create a
version2 of the function and keep the old symbol to not break
downstreams.

+++ libpkg/pkgvec.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 
 #define pkgvec_t(Type) \
-  struct { Type *d; int len, cap; }
+  struct { Type *d; size_t len, cap; }
 
 #define pkgvec_init(v) \
 	memset((v), 0, sizeof(*(v)))
